new ui contract where token is dynamic
